auth-service/requirements.txt
fastapi
uvicorn
pydantic
sqlalchemy
python-jose[cryptodigest]
bcrypt==4.1.3          # ← this version works perfectly with passlib
passlib[bcrypt]        # ← now safe
python-multipart
httpx
pyotp
qrcode[pil]